refactor(solicitacoes): Adicionando localStorageUtils
