+++ src/utils/localStorageUtils.js
@@ -1,0 +1,25 @@
+export const localStorageUtils = {
+    getItem(key) {
+        try {
+            const item = localStorage.getItem(key);
+            return item ? JSON.parse(item) : null;
+        } catch (error) {
+            console.error(`Error getting item from localStorage: ${key}`, error);
+            return null;
+        }
+    },
+    setItem(key, value) {
+        try {
+            localStorage.setItem(key, JSON.stringify(value));
+        } catch (error) {
+            console.error(`Error setting item in localStorage: ${key}`, error);
+        }
+    },
+    removeItem(key) {
+        try {
+            localStorage.removeItem(key);
+        } catch (error) {
+            console.error(`Error removing item from localStorage: ${key}`, error);
+        }
+    },
+};
